requirements.txt
numpy
opencv-python
tensorflow==1.14.0
matplotlib
pillow
